requirements.txt
google-auth-oauthlib==1.0
oauthlib==3.2.2
requests-oauthlib==1.3.1
google-analytics-data==0.18.0
google-cloud-core==2.3.3
google-cloud-bigquery==3.13.0
google-cloud-bigquery-storage==2.22.0
python-dateutil==2.8.2
google-cloud-secret-manager==2.16.3
pandas
facebook-business==16.0.0
json